+++ modules/router/src/test/services/transfer.spec.ts
@@ -28,20 +28,20 @@ import * as Sinon from "sinon";
 import { config } from "../../config";
 import { ForwardTransferError } from "../../errors";
 import { PrismaStore, RouterUpdateType } from "../../services/store";
-import { cancelCreatedTransfer, transferWithAutoCollateralization } from "../../services/transfer";
-import * as collateral from "../../collateral";
+import { cancelCreatedTransfer, attemptTransferWithCollateralization } from "../../services/transfer";
+import * as collateral from "../../services/collateral";
 
 const testName = "Router transfer service";
 
 const { log } = getTestLoggers(testName, config.logLevel as any);
 
 describe(testName, () => {
-  describe("transferWithAutoCollateralization", () => {
+  describe("attemptTransferWithCollateralization", () => {
     // Declare mocks
     let nodeService: Sinon.SinonStubbedInstance<RestServerNodeService>;
     let store: Sinon.SinonStubbedInstance<PrismaStore>;
     let chainReader: Sinon.SinonStubbedInstance<VectorChainReader>;
-    let requestCollateral: Sinon.SinonStub;
+    let justInTimeCollateral: Sinon.SinonStub;
 
     // Declare constants
     const routerPublicIdentifier = mkPublicIdentifier("vectorIII");
@@ -75,12 +75,12 @@ describe(testName, () => {
       nodeService = Sinon.createStubInstance(RestServerNodeService);
       store = Sinon.createStubInstance(PrismaStore);
       chainReader = Sinon.createStubInstance(VectorChainReader);
-      requestCollateral = Sinon.stub(collateral, "requestCollateral");
+      justInTimeCollateral = Sinon.stub(collateral, "justInTimeCollateral");
 
       // Default all stubs to return sucessful values
       nodeService.sendIsAliveMessage.resolves(Result.ok({ channelAddress }));
       store.queueUpdate.resolves(undefined);
-      requestCollateral.resolves(Result.ok(undefined));
+      justInTimeCollateral.resolves(Result.ok(undefined));
       nodeService.conditionalTransfer.resolves(Result.ok({ channelAddress, transferId, routingId }));
     });
 
@@ -101,7 +101,7 @@ describe(testName, () => {
       });
       nodeService.sendIsAliveMessage.resolves(Result.fail(new Error("fail") as any));
       const params = mkParams();
-      const res = await transferWithAutoCollateralization(
+      const res = await attemptTransferWithCollateralization(
         params,
         channel,
         routerPublicIdentifier,
@@ -127,7 +127,7 @@ describe(testName, () => {
         assetIds: [mkAddress()],
         balances: [{ to: [routerAddr, recipientAddr], amount: ["700", "53"] }],
       });
-      const res = await transferWithAutoCollateralization(
+      const res = await attemptTransferWithCollateralization(
         mkParams(),
         channel,
         routerPublicIdentifier,
@@ -141,7 +141,7 @@ describe(testName, () => {
       expect(res.getValue()).to.be.deep.eq({ channelAddress, transferId, routingId });
     });
 
-    it("should work if receiver is online and uncollateralized", async () => {
+    it("should work if receiver is online and undercollateralized", async () => {
       const { channel } = createTestChannelState(UpdateType.deposit, {
         channelAddress,
         aliceIdentifier: routerPublicIdentifier,
@@ -151,7 +151,7 @@ describe(testName, () => {
         assetIds: [mkAddress()],
         balances: [{ to: [routerAddr, recipientAddr], amount: ["0", "53"] }],
       });
-      const res = await transferWithAutoCollateralization(
+      const res = await attemptTransferWithCollateralization(
         mkParams(),
         channel,
         routerPublicIdentifier,
@@ -163,7 +163,7 @@ describe(testName, () => {
       );
       expect(res.getError()).to.be.undefined;
       expect(res.getValue()).to.be.deep.eq({ channelAddress, transferId, routingId });
-      expect(requestCollateral.callCount).to.be.eq(1);
+      expect(justInTimeCollateral.callCount).to.be.eq(1);
     });
 
     it("should fail if recipient if offline && requireOnline == true", async () => {
@@ -177,7 +177,7 @@ describe(testName, () => {
         balances: [{ to: [routerAddr, recipientAddr], amount: ["0", "53"] }],
       });
       nodeService.sendIsAliveMessage.resolves(Result.fail(new Error("fail") as any));
-      const res = await transferWithAutoCollateralization(
+      const res = await attemptTransferWithCollateralization(
         mkParams(),
         channel,
         routerPublicIdentifier,
@@ -202,7 +202,7 @@ describe(testName, () => {
         balances: [{ to: [routerAddr, recipientAddr], amount: ["700", "53"] }],
       });
       nodeService.sendIsAliveMessage.resolves(Result.fail(new Error("fail") as any));
-      const res = await transferWithAutoCollateralization(
+      const res = await attemptTransferWithCollateralization(
         mkParams(),
         channel,
         routerPublicIdentifier,
@@ -226,8 +226,8 @@ describe(testName, () => {
         assetIds: [mkAddress()],
         balances: [{ to: [routerAddr, recipientAddr], amount: ["0", "53"] }],
       });
-      requestCollateral.resolves(Result.fail(new Error("fail")));
-      const res = await transferWithAutoCollateralization(
+      justInTimeCollateral.resolves(Result.fail(new Error("fail")));
+      const res = await attemptTransferWithCollateralization(
         mkParams(),
         channel,
         routerPublicIdentifier,
@@ -252,7 +252,7 @@ describe(testName, () => {
         balances: [{ to: [routerAddr, recipientAddr], amount: ["0", "53"] }],
       });
       nodeService.conditionalTransfer.resolves(Result.fail(new Error("fail") as any));
-      const res = await transferWithAutoCollateralization(
+      const res = await attemptTransferWithCollateralization(
         mkParams(),
         channel,
         routerPublicIdentifier,