+++ modules/documentation/docs/changelog.md
@@ -2,7 +2,23 @@
 
 ## Next Release
 
-## 0.1.0-rc.3
+## 0.1.0-rc.9
+
+- \[router/server-node\] optionally skip checkin on start
+
+## 0.1.0-rc.8
+
+- \[messaging\] add `undefined` check on error handling
+
+## 0.1.0-rc.7
+
+- \[types\] relax `transferResolver` schema to allow for ethers decoded values
+
+## 0.1.0-rc.6
+
+- \[router\] fix prisma webpack
+
+## 0.1.0-rc.4
 
 - \[router\] add logdna service
 
@@ -13,6 +29,7 @@
 ## 0.1.0-rc.1
 
 - \[router\] add isAlive check before forwarding transfer to increase odds of successful transfer
+
 ## 0.1.0-rc.0
 
 - \[contracts\] audit results implemented